style: component 폴더 정리
