+++ src/components/homeCompo/HomeProfile.js
@@ -1,6 +1,6 @@
 import React from 'react';
 import { View, StyleSheet } from 'react-native';
-import { customTheme } from '../styles/customTheme.js';
+import { customTheme } from '../../styles/customTheme.js';
 
 const HomeProfile = () => {
   return (
