+++ src/components/icon_24/connect_df_24.js
