+++ src/components/icon_32/notification_32.js
